I changed the default font color in the Last Redoubt scheme to magenta, in the html element, to match the Darkening scheme. If any text appears magenta, its color is unset in the design, and should be set in some lower-ranking element.
I moved the story and related artwork copyright notices in the Darkening section to the same position it is in the Last Redoubt scheme: in the footer, above the breadcrumbs. It looks better.

I gave both the Last Redoubt and Darkening schemes general text colors for the footers.

I added examples of notes under the story-and-related-illos copyrights in the footers for both the Last Redoubt and the Darkening schemes. Occasionally desireable.